Ryddede op i projektet og flyttede funktioner ud i respektive mapper :P
