+++ cmake-build-debug/CMakeFiles/3.29.6/CMakeCCompiler.cmake
@@ -1,4 +1,4 @@
-set(CMAKE_C_COMPILER "D:/JetBrains/CLion 2024.2.2/bin/mingw/bin/gcc.exe")
+set(CMAKE_C_COMPILER "D:/CLion 2024.2/bin/mingw/bin/gcc.exe")
 set(CMAKE_C_COMPILER_ARG1 "")
 set(CMAKE_C_COMPILER_ID "GNU")
 set(CMAKE_C_COMPILER_VERSION "13.1.0")
@@ -21,14 +21,14 @@ set(CMAKE_C_SIMULATE_VERSION "")
 
 
 
-set(CMAKE_AR "D:/JetBrains/CLion 2024.2.2/bin/mingw/bin/ar.exe")
-set(CMAKE_C_COMPILER_AR "D:/JetBrains/CLion 2024.2.2/bin/mingw/bin/gcc-ar.exe")
-set(CMAKE_RANLIB "D:/JetBrains/CLion 2024.2.2/bin/mingw/bin/ranlib.exe")
-set(CMAKE_C_COMPILER_RANLIB "D:/JetBrains/CLion 2024.2.2/bin/mingw/bin/gcc-ranlib.exe")
-set(CMAKE_LINKER "D:/JetBrains/CLion 2024.2.2/bin/mingw/bin/ld.exe")
+set(CMAKE_AR "D:/CLion 2024.2/bin/mingw/bin/ar.exe")
+set(CMAKE_C_COMPILER_AR "D:/CLion 2024.2/bin/mingw/bin/gcc-ar.exe")
+set(CMAKE_RANLIB "D:/CLion 2024.2/bin/mingw/bin/ranlib.exe")
+set(CMAKE_C_COMPILER_RANLIB "D:/CLion 2024.2/bin/mingw/bin/gcc-ranlib.exe")
+set(CMAKE_LINKER "D:/CLion 2024.2/bin/mingw/bin/ld.exe")
 set(CMAKE_LINKER_LINK "")
 set(CMAKE_LINKER_LLD "")
-set(CMAKE_C_COMPILER_LINKER "D:/JetBrains/CLion 2024.2.2/bin/mingw/bin/ld.exe")
+set(CMAKE_C_COMPILER_LINKER "D:/CLion 2024.2/bin/mingw/bin/ld.exe")
 set(CMAKE_C_COMPILER_LINKER_ID "GNU")
 set(CMAKE_C_COMPILER_LINKER_VERSION 2.40)
 set(CMAKE_C_COMPILER_LINKER_FRONTEND_VARIANT GNU)
@@ -74,7 +74,7 @@ endif()
 
 
 
-set(CMAKE_C_IMPLICIT_INCLUDE_DIRECTORIES "D:/JetBrains/CLion 2024.2.2/bin/mingw/lib/gcc/x86_64-w64-mingw32/13.1.0/include;D:/JetBrains/CLion 2024.2.2/bin/mingw/include;D:/JetBrains/CLion 2024.2.2/bin/mingw/lib/gcc/x86_64-w64-mingw32/13.1.0/include-fixed;D:/JetBrains/CLion 2024.2.2/bin/mingw/x86_64-w64-mingw32/include")
+set(CMAKE_C_IMPLICIT_INCLUDE_DIRECTORIES "D:/CLion 2024.2/bin/mingw/lib/gcc/x86_64-w64-mingw32/13.1.0/include;D:/CLion 2024.2/bin/mingw/include;D:/CLion 2024.2/bin/mingw/lib/gcc/x86_64-w64-mingw32/13.1.0/include-fixed;D:/CLion 2024.2/bin/mingw/x86_64-w64-mingw32/include")
 set(CMAKE_C_IMPLICIT_LINK_LIBRARIES "")
 set(CMAKE_C_IMPLICIT_LINK_DIRECTORIES "")
 set(CMAKE_C_IMPLICIT_LINK_FRAMEWORK_DIRECTORIES "")
