+++ cmake-build-debug/CMakeFiles/3.29.6/CMakeRCCompiler.cmake
@@ -1,4 +1,4 @@
-set(CMAKE_RC_COMPILER "D:/JetBrains/CLion 2024.2.2/bin/mingw/bin/windres.exe")
+set(CMAKE_RC_COMPILER "D:/CLion 2024.2/bin/mingw/bin/windres.exe")
 set(CMAKE_RC_COMPILER_ARG1 "")
 set(CMAKE_RC_COMPILER_LOADED 1)
 set(CMAKE_RC_SOURCE_FILE_EXTENSIONS rc;RC)
