+++ cmake-build-debug/CMakeFiles/clion-Debug-log.txt
@@ -1,10 +1,4 @@
-"D:\JetBrains\CLion 2024.2.2\bin\cmake\win\x64\bin\cmake.exe" -DCMAKE_BUILD_TYPE=Debug -G "MinGW Makefiles" -S D:\AAU\IMPR\P1_Minefield -B D:\AAU\IMPR\P1_Minefield\cmake-build-debug
--- The C compiler identification is GNU 13.1.0
--- Detecting C compiler ABI info
--- Detecting C compiler ABI info - done
--- Check for working C compiler: D:/JetBrains/CLion 2024.2.2/bin/mingw/bin/gcc.exe - skipped
--- Detecting C compile features
--- Detecting C compile features - done
--- Configuring done (0.8s)
--- Generating done (0.0s)
--- Build files have been written to: D:/AAU/IMPR/P1_Minefield/cmake-build-debug
+"D:\CLion 2024.2\bin\cmake\win\x64\bin\cmake.exe" -DCMAKE_BUILD_TYPE=Debug -G "MinGW Makefiles" -S C:\Users\kaspe\Documents\GitHub\P1_Minefield -B C:\Users\kaspe\Documents\GitHub\P1_Minefield\cmake-build-debug
+-- Configuring done (0.2s)
+-- Generating done (0.2s)
+-- Build files have been written to: C:/Users/kaspe/Documents/GitHub/P1_Minefield/cmake-build-debug
