+++ cmake-build-debug/CMakeFiles/clion-environment.txt
@@ -1,5 +1,5 @@
-ToolSet: 11.0 w64 (local)@D:\JetBrains\CLion 2024.2.2\bin\mingw
-Ninja: 1.12.0@D:\JetBrains\CLion 2024.2.2\bin\ninja\win\x64\ninja.exe
+ToolSet: 11.0 w64 (local)@D:\CLion 2024.2\bin\mingw
+Ninja: 1.12.0@D:\CLion 2024.2\bin\ninja\win\x64\ninja.exe
 Options: 
 
 Options:
