+++ cmake-build-debug/Testing/Temporary/LastTest.log
@@ -1,3 +1,3 @@
-Start testing: Dec 05 11:14 Rom, normaltid
+Start testing: Dec 05 11:46 Rom, normaltid
 ----------------------------------------------------------
-End testing: Dec 05 11:14 Rom, normaltid
+End testing: Dec 05 11:46 Rom, normaltid
