+++ src/blast_radius/blast_radius.h
@@ -1,0 +1,11 @@
+//
+// Created by kaspe on 05-12-2024.
+//
+
+
+#ifndef BLAST_RADIUS_H
+#define BLAST_RADIUS_H
+#include "src/map/map.h"
+void blast_radius(int mapSize, mapPoint* map);
+
+#endif //BLAST_RADIUS_H
