+++ src/map/map.h
@@ -4,7 +4,9 @@
 #define MAP_H
 #include "src/structs.h"
 #include "src/map/colour/colour.h"
-
+#include "src/path/start_point/start_point.h"
+#include <stdio.h>
+#include <stdlib.h>
 
 mapPoint* getCell(mapPoint* map, int mapSize, int y, int x);
 
@@ -13,10 +15,7 @@ void create_map(int mapSize, mapPoint* map, int* amount_of_mines);
 void print_map(int mapSize, mapPoint* map, Deminer* deminers, int amount_of_deminers);
 void writeToMap(char ** map, int * map_size);
 void check_a_point(int mapSize, mapPoint* map);
-void red();
-void green();
-void reset();
-void yellow();
+void print_map_info(int mapSize, mapPoint* map);
